add examples for *.map class method

+++ app/models/category_mapping.rb
@@ -13,13 +13,18 @@ class CategoryMapping < ActiveRecord::Base
 
   def self.map(question, title)
     category = Category.where(title: title).first
-    category_mappings = CategoryMapping.where(category_id: category.id, question_id: question.id)
-    category_mappings.first || CategoryMapping.create(category_id: category.id, question_id: question.id)
+    if category
+      category_mappings = CategoryMapping.where(category_id: category.id, question_id: question.id)
+      category_mappings.first || CategoryMapping.create(category_id: category.id, question_id: question.id)
+    end
   end
 
   def self.unmap(question, title)
     category = Category.where(title: title).first
-    category_mappings = CategoryMapping.where(category_id: category.id, question_id: question.id)
-    category_mappings.destroy_all
+    if category
+      category_mappings = CategoryMapping.where(category_id: category.id, question_id: question.id)
+      category_mappings.destroy_all
+    end
   end
+
 end
